API: Split callbacks modules and make matplotlib optional.
- Move plotting callbacks into bluesky.callbacks.mpl_plotting
- Shim import them into bluesky.callbacks for back-compat (only if matplotlib
  is importable).
- Move LiveFit and PeakStats into a new module, bluesky.callbacks.fitting.
- Remove the module bluesky.callbacks.scientific, which previously only held
  PeakStats and was not well named.
- Remove module-level import of matplotlib in bluesky.simulators.
- Remove matplotlib from requirements.txt. It is already in
  test-requirements.txt.
