+++ bluesky/callbacks/__init__.py
@@ -1,4 +1,10 @@
 from .core import (CallbackBase, CallbackCounter, print_metadata, collector,
-                   LiveScatter, LivePlot, LiveGrid, LiveTable, LiveFit,
-                   LiveFitPlot, CollectThenCompute, _get_obj_fields,
-                   LiveRaster, LiveMesh)
+                   get_obj_fields, CollectThenCompute, LiveTable)
+from .fitting import LiveFit
+try:
+    import matplotlib
+except ImportError:
+    ...
+else:
+    from .mpl_plotting import (LiveScatter, LivePlot, LiveGrid,
+                               LiveFitPlot, LiveRaster, LiveMesh)
