+++ requirements.txt
@@ -2,7 +2,6 @@ cycler
 event-model
 historydict
 jsonschema
-matplotlib
 numpy
 super_state_machine
 toolz
